# backend/src/swagger.yaml
openapi: "3.0.3"
info:
  title: HalalChain V2 API
  description: Enterprise Halal Supply Chain Platform — REST API
  version: "2.0.0"
  contact:
    name: HalalChain Team
servers:
  - url: http://localhost:4000
    description: Local development server
  - url: https://api.halalchain.app
    description: Production server

# ─────────────────────────────────────────────────────────────
# Security
# ─────────────────────────────────────────────────────────────
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token

  # ─────────────────────────────────────────────────────────────
  # Schemas
  # ─────────────────────────────────────────────────────────────
  schemas:
...
    PurchaseOrder:
      type: object
      required:
        - id
        - supplierId
        - poNumber
        - status
        - totalAmount
        - createdAt
      properties:
        id:
          type: string
          format: uuid
          example: "cc0e8400-e29b-41d4-a716-446655440007"
        supplierId:
          type: string
          format: uuid
          example: "660e8400-e29b-41d4-a716-446655440001"
        poNumber:
          type: string
          example: "PO-2024-00042"
        status:
          type: string
          enum:
            - DRAFT
            - APPROVED
            - SHIPPING
            - RECEIVED
            - CANCELLED
            - PARTIAL
          example: "APPROVED"
        totalAmount:
          type: number
          format: float
          example: 3750.00
        createdAt:
          type: string
          format: date-time
          example: "2024-05-20T11:00:00.000Z"
        itemCount:
          type: integer
          description: Number of line items in the purchase order (optional — present in list responses)
          example: 4

    # ── PurchaseOrderItem ──────────────────────────────────────
    PurchaseOrderItem:
      type: object
      required:
        - id
        - purchaseOrderId
        - productId
        - quantity
        - unitPrice
        - subtotal
        - createdAt
      properties:
        id:
          type: string
          format: uuid
          example: "dd0e8400-e29b-41d4-a716-446655440008"
        purchaseOrderId:
          type: string
          format: uuid
          example: "cc0e8400-e29b-41d4-a716-446655440007"
        productId:
          type: string
          format: uuid
          example: "770e8400-e29b-41d4-a716-446655440002"
        quantity:
          type: integer
          minimum: 1
          example: 300
        unitPrice:
          type: number
          format: float
          minimum: 0.01
          example: 12.50
        subtotal:
          type: number
          format: float
          example: 3750.00
        createdAt:
          type: string
          format: date-time
          example: "2024-05-20T11:00:00.000Z"

    # ── Shipment ───────────────────────────────────────────────
...
    PaginatedResponse:
      type: object
      required:
        - data
        - page
        - limit
        - total
        - totalPages
      properties:
        data:
          type: array
          items:
            type: object
          description: Array of result items for the current page
        page:
          type: integer
          minimum: 1
          example: 1
        limit:
          type: integer
          minimum: 1
          maximum: 100
          example: 20
        total:
          type: integer
          minimum: 0
          example: 85
        totalPages:
          type: integer
          minimum: 0
          example: 5

    # ── ErrorResponse ──────────────────────────────────────────
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          example: "Resource not found"

# ─────────────────────────────────────────────────────────────
# Paths
# ─────────────────────────────────────────────────────────────
paths:
...
  /api/purchase-orders:
    get:
      tags:
        - Purchase Orders
      summary: List purchase orders (paginated, filterable)
      description: Returns a paginated list of purchase orders. Supports filtering by status and supplierId. Requires ADMIN or MANAGER role.
      security:
        - cookieAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: status
          in: query
          description: Filter by PO status
          schema:
            type: string
            enum:
              - DRAFT
              - APPROVED
              - SHIPPING
              - RECEIVED
              - CANCELLED
              - PARTIAL
            example: "APPROVED"
        - name: supplierId
          in: query
          description: Filter by supplier UUID; returns empty list if supplier does not exist
          schema:
            type: string
            format: uuid
            example: "660e8400-e29b-41d4-a716-446655440001"
      responses:
        "200":
          description: Paginated list of purchase orders (each includes `itemCount`)
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/PaginatedResponse"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/PurchaseOrder"
        "400":
          description: Invalid status value
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "Invalid status value."
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Insufficient role
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    post:
      tags:
        - Purchase Orders
      summary: Create a new purchase order (ADMIN only)
      description: Creates a PO with one or more line items. Server computes `subtotal` per item and `totalAmount`. A `poNumber` is auto-generated if not provided.
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - supplierId
                - items
              properties:
                supplierId:
                  type: string
                  format: uuid
                  example: "660e8400-e29b-41d4-a716-446655440001"
                poNumber:
                  type: string
                  minLength: 3
                  maxLength: 50
                  description: Optional; auto-generated as PO-YYYY-NNNN if omitted
                  example: "PO-2024-00042"
                items:
                  type: array
                  minItems: 1
                  items:
                    type: object
                    required:
                      - productId
                      - quantity
                      - unitPrice
                    properties:
                      productId:
                        type: string
                        format: uuid
                        example: "770e8400-e29b-41d4-a716-446655440002"
                      quantity:
                        type: integer
                        minimum: 1
                        example: 300
                      unitPrice:
                        type: number
                        minimum: 0.01
                        example: 12.50
            example:
              supplierId: "660e8400-e29b-41d4-a716-446655440001"
              items:
                - productId: "770e8400-e29b-41d4-a716-446655440002"
                  quantity: 300
                  unitPrice: 12.50
                - productId: "771e8400-e29b-41d4-a716-446655440003"
                  quantity: 50
                  unitPrice: 8.00
      responses:
        "201":
          description: Purchase order created with all line items
          content:
            application/json:
              schema:
                type: object
                properties:
                  purchaseOrder:
                    allOf:
                      - $ref: "#/components/schemas/PurchaseOrder"
                      - type: object
                        properties:
                          items:
                            type: array
                            items:
                              $ref: "#/components/schemas/PurchaseOrderItem"
              example:
                purchaseOrder:
                  id: "cc0e8400-e29b-41d4-a716-446655440007"
                  supplierId: "660e8400-e29b-41d4-a716-446655440001"
                  poNumber: "PO-2024-00042"
                  status: "DRAFT"
                  totalAmount: 4150.00
                  createdAt: "2024-05-20T11:00:00.000Z"
                  supplier:
                    name: "Al-Barakah Foods Sdn Bhd"
                    country: "Malaysia"
                  items:
                    - id: "dd0e8400-e29b-41d4-a716-446655440008"
                      purchaseOrderId: "cc0e8400-e29b-41d4-a716-446655440007"
                      productId: "770e8400-e29b-41d4-a716-446655440002"
                      quantity: 300
                      unitPrice: 12.50
                      subtotal: 3750.00
                      createdAt: "2024-05-20T11:00:00.000Z"
                    - id: "dd1e8400-e29b-41d4-a716-446655440009"
                      purchaseOrderId: "cc0e8400-e29b-41d4-a716-446655440007"
                      productId: "771e8400-e29b-41d4-a716-446655440003"
                      quantity: 50
                      unitPrice: 8.00
                      subtotal: 400.00
                      createdAt: "2024-05-20T11:00:00.000Z"
        "400":
          description: Validation error — missing or invalid items
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "Invalid items"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Insufficient role
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Supplier or product not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "Product not found"
  /api/purchase-orders/{id}:
    get:
      tags:
        - Purchase Orders
      summary: Get a single purchase order by ID (includes items)
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Purchase order with full item details
          content:
            application/json:
              schema:
                type: object
                properties:
                  purchaseOrder:
                    allOf:
                      - $ref: "#/components/schemas/PurchaseOrder"
                      - type: object
                        properties:
                          items:
                            type: array
                            items:
                              $ref: "#/components/schemas/PurchaseOrderItem"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Insufficient role
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Purchase order not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      tags:
        - Purchase Orders
      summary: Delete a purchase order (ADMIN only, DRAFT status only)
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Purchase order deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Purchase order deleted"
        "400":
          description: Cannot delete — not in DRAFT status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Insufficient role
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Purchase order not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/purchase-orders/{id}/status:
    patch:
      tags:
        - Purchase Orders
      summary: Update purchase order status (ADMIN only)
      description: Transitions the PO status following allowed progression — DRAFT→APPROVED→SHIPPING→RECEIVED.
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - status
              properties:
                status:
                  type: string
                  enum:
                    - DRAFT
                    - APPROVED
                    - SHIPPING
                    - RECEIVED
                    - CANCELLED
                    - PARTIAL
                  example: "APPROVED"
      responses:
        "200":
          description: Status updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  purchaseOrder:
                    $ref: "#/components/schemas/PurchaseOrder"
        "400":
          description: Invalid status transition or validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Insufficient role
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Purchase order not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ══════════════════════════════════════════════════════════
  # Shipments
  # ══════════════════════════════════════════════════════════
...